+++ ventas/models.py
@@ -9,7 +9,7 @@ from servicios.models import Servicio
 # Create your models here.
 class DetalleVenta(models.Model):
     cantidadProducto = models.BigIntegerField(validators = [ MinValueValidator ( 0 )], verbose_name="Cantidad Producto") 
-    valorTotalProducto = models.PositiveBigIntegerField(validators = [ MinValueValidator ( 0 )], verbose_name="Valor Total Producto")
+    valorTotalProducto = models.PositiveIntegerField(validators = [ MinValueValidator ( 0 )], verbose_name="Valor Total Producto")
     producto = models.ForeignKey(Producto, on_delete=models.CASCADE, verbose_name="Producto",null=True)
     
     class Meta:
@@ -18,10 +18,10 @@ class DetalleVenta(models.Model):
 
 
 class Venta(models.Model):
-    subTotalVenta= models.PositiveBigIntegerField(validators = [ MinValueValidator ( 0 )], verbose_name="Subtotal Venta")
+    subTotalVenta= models.PositiveIntegerField(validators = [ MinValueValidator ( 0 )], verbose_name="Subtotal Venta")
     fecha= models.DateTimeField(verbose_name="Fecha Venta",help_text= "MM/DD/AAAA")
     porcentajeIva=models.DecimalField(validators=[MinValueValidator(0.0)],decimal_places=1,max_digits=20, verbose_name="Porcentaje IVA")
-    totalVenta= models.PositiveBigIntegerField(validators = [ MinValueValidator ( 0 )], verbose_name="Total Venta")
+    totalVenta= models.PositiveIntegerField(validators = [ MinValueValidator ( 0 )], verbose_name="Total Venta")
     detalleVenta= models.ForeignKey(DetalleVenta, on_delete=models.CASCADE, verbose_name="Detalle Venta",null=True)
     cliente=models.ForeignKey(Cliente, on_delete=models.CASCADE, verbose_name="Cliente",null=True)
     usuario=models.ForeignKey(Usuario, on_delete=models.CASCADE, verbose_name="Usuario",null=True)
